+++ src/main/java/net/sf/mpxj/CustomFieldContainer.java
@@ -98,9 +98,9 @@ public class CustomFieldContainer implements Iterable<CustomField>
    public void registerValue(CustomFieldValueItem item)
    {
       m_valueMap.put(item.getUniqueID(), item);
-      if (item.getGuid() != null)
+      if (item.getGUID() != null)
       {
-         m_guidMap.put(item.getGuid(), item);
+         m_guidMap.put(item.getGUID(), item);
       }
    }
 
@@ -112,9 +112,9 @@ public class CustomFieldContainer implements Iterable<CustomField>
    public void deregisterValue(CustomFieldValueItem item)
    {
       m_valueMap.remove(item.getUniqueID());
-      if (item.getGuid() != null)
+      if (item.getGUID() != null)
       {
-         m_guidMap.remove(item.getGuid());
+         m_guidMap.remove(item.getGUID());
       }
    }