+++ src/main/java/net/sf/mpxj/CustomFieldValueDataType.java
@@ -30,23 +30,27 @@ import net.sf.mpxj.common.EnumHelper;
  */
 public enum CustomFieldValueDataType implements MpxjEnum
 {
-   DATE(4),
-   DURATION(6), 
-   COST(9),
-   NUMBER(15),
-   FLAG(17),
-   TEXT(21),
-   FINISH_DATE(27);
+   DATE(4, 4, DataType.DATE),
+   DURATION(6, 6, DataType.DURATION), 
+   COST(9, 5, DataType.CURRENCY),
+   NUMBER(15, 7, DataType.NUMERIC),
+   FLAG(17, 17, DataType.BOOLEAN),
+   TEXT(21, 3, DataType.STRING),
+   FINISH_DATE(27, 9, DataType.DATE);
 
    
    /**
     * Private constructor.
     *
-    * @param type int version of the enum
+    * @param value int version of the enum
+    * @param maskValue data type used in mask definition 
+    * @param type data type
     */
-   private CustomFieldValueDataType(int type)
+   private CustomFieldValueDataType(int value, int maskValue, DataType type)
    {
-      m_value = type;
+      m_value = value;
+      m_maskValue = maskValue;
+      m_type = type;
    }
 
    /**
@@ -65,6 +69,21 @@ public enum CustomFieldValueDataType implements MpxjEnum
    }
 
    /**
+    * Retrieve an instance of the enum based on its mask value.
+    *
+    * @param type mask value
+    * @return enum instance
+    */
+   public static CustomFieldValueDataType getInstanceByMaskValue(int type)
+   {
+      if (type >= 0 && type < MASK_VALUES.length)
+      {
+         return MASK_VALUES[type];
+      }
+      return null;
+   }
+
+   /**
     * Accessor method used to retrieve the numeric representation of the enum.
     *
     * @return int representation of the enum
@@ -75,12 +94,45 @@ public enum CustomFieldValueDataType implements MpxjEnum
    }
 
    /**
+    * Retrieve the MPXJ data type.
+    * 
+    * @return MPXJ data type
+    */
+   public DataType getDataType()
+   {
+      return m_type;
+   }
+   
+   /**
+    * Retrieve the mask value.
+    * 
+    * @return mask value
+    */
+   public int getMaskValue()
+   {
+      return m_maskValue;
+   }
+
+   /**
     * Array mapping int types to enums.
     */
    private static final CustomFieldValueDataType[] TYPE_VALUES = EnumHelper.createTypeArray(CustomFieldValueDataType.class, 21);
 
+   private static final CustomFieldValueDataType[] MASK_VALUES = new CustomFieldValueDataType[28];
+   static
+   {
+      for (CustomFieldValueDataType value : values())
+      {
+         MASK_VALUES[value.getMaskValue()] = value;
+      }
+   }
+   
    /**
     * Internal representation of the enum int type.
     */
-   private int m_value;
+   private final int m_value;
+   
+   private final int m_maskValue;
+   
+   private final DataType m_type;
 }