+++ src/main/java/net/sf/mpxj/mpp/CustomFieldValueReader14.java
@@ -95,7 +95,7 @@ public class CustomFieldValueReader14 extends CustomFieldValueReader
          byte[] b2 = m_outlineCodeFixedData2.getByteArrayValue(loop + 3);
          if (b2 != null)
          {
-            item.setGuid(MPPUtility.getGUID(b2, 0));
+            item.setGUID(MPPUtility.getGUID(b2, 0));
             UUID lookupTableGuid = MPPUtility.getGUID(b2, fieldOffset);
             item.setType(CustomFieldValueDataType.getInstance(MPPUtility.getShort(b2, typeOffset)));
             item.setValue(getTypedValue(item.getType(), value));            
@@ -131,13 +131,9 @@ public class CustomFieldValueReader14 extends CustomFieldValueReader
          {
             int extendedAttributeFieldID = MPPUtility.getInt(data, index + 0);
             FieldType field = FieldTypeHelper.getInstance(extendedAttributeFieldID);
-            if (field != null)
-            {
-               // UUID extendedAttributeGuid = MPPUtility.getGUID(data, index + 20);
-               UUID lookupTableGuid = MPPUtility.getGUID(data, index + 36);
-               map.put(lookupTableGuid, field);
-            }
-
+            // UUID extendedAttributeGuid = MPPUtility.getGUID(data, index + 20);
+            UUID lookupTableGuid = MPPUtility.getGUID(data, index + 36);
+            map.put(lookupTableGuid, field);
             index += 88;
          }
       }