+++ src/main/java/net/sf/mpxj/mspdi/MSPDIReader.java
@@ -34,7 +34,9 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
@@ -56,6 +58,10 @@ import net.sf.mpxj.Availability;
 import net.sf.mpxj.AvailabilityTable;
 import net.sf.mpxj.CostRateTable;
 import net.sf.mpxj.CostRateTableEntry;
+import net.sf.mpxj.CustomField;
+import net.sf.mpxj.CustomFieldLookupTable;
+import net.sf.mpxj.CustomFieldValueDataType;
+import net.sf.mpxj.CustomFieldValueMask;
 import net.sf.mpxj.DateRange;
 import net.sf.mpxj.Day;
 import net.sf.mpxj.DayType;
@@ -99,6 +105,7 @@ import net.sf.mpxj.common.Pair;
 import net.sf.mpxj.common.SplitTaskFactory;
 import net.sf.mpxj.common.TimephasedWorkNormaliser;
 import net.sf.mpxj.listener.ProjectListener;
+import net.sf.mpxj.mpp.CustomFieldValueItem;
 import net.sf.mpxj.mspdi.schema.Project;
 import net.sf.mpxj.mspdi.schema.Project.Calendars.Calendar.WorkWeeks;
 import net.sf.mpxj.mspdi.schema.Project.Calendars.Calendar.WorkWeeks.WorkWeek;
@@ -230,6 +237,7 @@ public final class MSPDIReader extends AbstractProjectReader
          readResources(project, calendarMap);
          readTasks(project);
          readAssignments(project);
+         readOutlineCodes(project);
 
          //
          // Ensure that the unique ID counters are correct
@@ -271,6 +279,7 @@ public final class MSPDIReader extends AbstractProjectReader
       finally
       {
          m_projectFile = null;
+         m_lookupTableMap.clear();
       }
    }
 
@@ -844,10 +853,11 @@ public final class MSPDIReader extends AbstractProjectReader
     */
    private void readFieldAlias(Project.ExtendedAttributes.ExtendedAttribute attribute)
    {
+      FieldType field = FieldTypeHelper.getInstance(Integer.parseInt(attribute.getFieldID()));
+      m_lookupTableMap.put(attribute.getLtuid(), field);
       String alias = attribute.getAlias();
       if (alias != null && alias.length() != 0)
       {
-         FieldType field = FieldTypeHelper.getInstance(Integer.parseInt(attribute.getFieldID()));
          m_projectFile.getCustomFields().getCustomField(field).setAlias(attribute.getAlias());
       }
    }
@@ -1346,7 +1356,7 @@ public final class MSPDIReader extends AbstractProjectReader
          {
             mpx.set(TaskField.ACTUAL_DURATION_UNITS, mpx.getActualDuration().getUnits());
          }
-         
+
          //
          // When reading an MSPDI file, the project summary task contains
          // some of the values used to populate the project properties.
@@ -1566,6 +1576,118 @@ public final class MSPDIReader extends AbstractProjectReader
          for (Project.Assignments.Assignment assignment : assignments.getAssignment())
          {
             readAssignment(assignment, splitFactory, normaliser);
+         }
+      }
+   }
+
+   /**
+    * This method extracts outline code/custom field data from an MSPDI file.
+    * 
+    * @param project Root node of the MSPDI file
+    */
+   private void readOutlineCodes(Project project)
+   {
+      Project.OutlineCodes outlineCodes = project.getOutlineCodes();
+      if (outlineCodes != null)
+      {
+         for (Project.OutlineCodes.OutlineCode outlineCode : outlineCodes.getOutlineCode())
+         {
+            readOutlineCode(outlineCode);
+         }
+      }
+   }
+
+   /**
+    * This method extracts the definition of a single outline code/custom lookup table from an MSPDI file.
+    * 
+    * @param outlineCode outline code data from the MSPDI file
+    */
+   private void readOutlineCode(Project.OutlineCodes.OutlineCode outlineCode)
+   {
+      FieldType fieldType;
+      String fieldID = outlineCode.getFieldID();
+      if (fieldID == null)
+      {
+         fieldType = m_lookupTableMap.get(outlineCode.getGuid());
+      }
+      else
+      {
+         fieldType = FieldTypeHelper.getInstance(NumberHelper.getInt(outlineCode.getFieldID()));
+      }
+
+      if (fieldType != null)
+      {
+         CustomField field = m_projectFile.getCustomFields().getCustomField(fieldType);
+         String currentAlias = field.getAlias();
+         // Don't overwrite an alias we've read from extended attributes
+         if (currentAlias == null || currentAlias.isEmpty())
+         {
+            field.setAlias(outlineCode.getAlias());   
+         }         
+         readOutlineCodeValues(outlineCode, field);
+         readOutlineCodeMasks(outlineCode, field);
+      }
+   }
+
+   /**
+    * This method extracts the lookup table values for an outline code/custom field from an MSPDI file.
+    * 
+    * @param outlineCode outline code data from the MSPDI file
+    * @param field target field
+    */
+   private void readOutlineCodeValues(Project.OutlineCodes.OutlineCode outlineCode, CustomField field)
+   {
+      Project.OutlineCodes.OutlineCode.Values values = outlineCode.getValues();
+      if (values != null)
+      {
+         CustomFieldLookupTable table = field.getLookupTable();
+
+         table.setEnterprise(BooleanHelper.getBoolean(outlineCode.isEnterprise()));
+         table.setAllLevelsRequired(BooleanHelper.getBoolean(outlineCode.isAllLevelsRequired()));
+         table.setGUID(outlineCode.getGuid());
+         table.setLeafOnly(BooleanHelper.getBoolean(outlineCode.isLeafOnly()));
+         table.setOnlyTableValuesAllowed(BooleanHelper.getBoolean(outlineCode.isOnlyTableValuesAllowed()));
+         table.setResourceSubstitutionEnabled(BooleanHelper.getBoolean(outlineCode.isResourceSubstitutionEnabled()));
+         table.setShowIndent(BooleanHelper.getBoolean(outlineCode.isShowIndent()));
+
+         for (Project.OutlineCodes.OutlineCode.Values.Value value : values.getValue())
+         {
+            CustomFieldValueItem item = new CustomFieldValueItem(NumberHelper.getInteger(value.getValueID()));
+            item.setDescription(value.getDescription());
+            item.setGUID(value.getFieldGUID());
+            item.setCollapsed(BooleanHelper.getBoolean(value.isIsCollapsed()));
+            item.setParent(NumberHelper.getInteger(value.getParentValueID()));
+            item.setType(CustomFieldValueDataType.getInstance(NumberHelper.getInt(value.getType())));
+            item.setValue(DatatypeConverter.parseOutlineCodeValue(value.getValue(), field.getFieldType().getDataType()));
+            table.add(item);
+         }
+      }
+   }
+
+   /**
+    * This method extracts the lookup table masks for an outline code/custom field from an MSPDI file.
+    * 
+    * @param outlineCode outline code data from the MSPDI file
+    * @param field target field
+    */
+   private void readOutlineCodeMasks(Project.OutlineCodes.OutlineCode outlineCode, CustomField field)
+   {
+      Project.OutlineCodes.OutlineCode.Masks masks = outlineCode.getMasks();
+      if (masks != null)
+      {
+         List<CustomFieldValueMask> maskList = field.getMasks();
+         for (Project.OutlineCodes.OutlineCode.Masks.Mask mask : masks.getMask())
+         {
+            int length = NumberHelper.getInt(mask.getLength());
+            int level = NumberHelper.getInt(mask.getLevel());
+            String separator = mask.getSeparator();
+            CustomFieldValueDataType type = CustomFieldValueDataType.getInstanceByMaskValue(NumberHelper.getInt(mask.getType()));
+            if (type == null)
+            {
+               type = CustomFieldValueDataType.TEXT;
+            }
+            CustomFieldValueMask item = new CustomFieldValueMask(length, level, separator, type);
+            maskList.add(item);
          }
       }
    }
@@ -1874,6 +1996,7 @@ public final class MSPDIReader extends AbstractProjectReader
    private ProjectFile m_projectFile;
    private EventManager m_eventManager;
    private List<ProjectListener> m_projectListeners;
+   private Map<UUID, FieldType> m_lookupTableMap = new HashMap<UUID, FieldType>();
 
    private static final RecurrenceType[] RECURRENCE_TYPES =
    {