+++ src/main/java/net/sf/mpxj/mspdi/MSPDIWriter.java
@@ -48,6 +48,9 @@ import net.sf.mpxj.CostRateTable;
 import net.sf.mpxj.CostRateTableEntry;
 import net.sf.mpxj.CustomField;
 import net.sf.mpxj.CustomFieldContainer;
+import net.sf.mpxj.CustomFieldLookupTable;
+import net.sf.mpxj.CustomFieldValueDataType;
+import net.sf.mpxj.CustomFieldValueMask;
 import net.sf.mpxj.DateRange;
 import net.sf.mpxj.Day;
 import net.sf.mpxj.DayType;
@@ -83,6 +86,7 @@ import net.sf.mpxj.common.MPPTaskField;
 import net.sf.mpxj.common.NumberHelper;
 import net.sf.mpxj.common.ResourceFieldLists;
 import net.sf.mpxj.common.TaskFieldLists;
+import net.sf.mpxj.mpp.CustomFieldValueItem;
 import net.sf.mpxj.mspdi.schema.ObjectFactory;
 import net.sf.mpxj.mspdi.schema.Project;
 import net.sf.mpxj.mspdi.schema.Project.Calendars.Calendar.Exceptions;
@@ -195,6 +199,7 @@ public final class MSPDIWriter extends AbstractProjectWriter
          writeResources(project);
          writeTasks(project);
          writeAssignments(project);
+         writeOutlineCodes(project);
          writeProjectExtendedAttributes(project);
 
          marshaller.marshal(project, stream);
@@ -309,13 +314,12 @@ public final class MSPDIWriter extends AbstractProjectWriter
       }
 
       customFields.addAll(m_extendedAttributesInUse);
-      
+
       List<FieldType> customFieldsList = new ArrayList<FieldType>();
       customFieldsList.addAll(customFields);
-      
 
       // Sort to ensure consistent order in file
-      final CustomFieldContainer customFieldContainer =  m_projectFile.getCustomFields();
+      final CustomFieldContainer customFieldContainer = m_projectFile.getCustomFields();
       Collections.sort(customFieldsList, new Comparator<FieldType>()
       {
          @Override public int compare(FieldType o1, FieldType o2)
@@ -337,7 +341,150 @@ public final class MSPDIWriter extends AbstractProjectWriter
 
          CustomField customField = customFieldContainer.getCustomField(fieldType);
          attribute.setAlias(customField.getAlias());
+         attribute.setLtuid(customField.getLookupTable().getGUID());
       }
+   }
+
+   /**
+    * Write outline code/custom field lookup tables.
+    * 
+    * @param project Root node of the MSPDI file
+    */
+   private void writeOutlineCodes(Project project)
+   {
+      Project.OutlineCodes outlineCodes = null;
+      List<CustomField> allCustomFields = new ArrayList<CustomField>();
+      for (CustomField field : m_projectFile.getCustomFields())
+      {
+         allCustomFields.add(field);
+      }
+
+      Collections.sort(allCustomFields, new Comparator<CustomField>()
+      {
+         @Override public int compare(CustomField customField1, CustomField customField2)
+         {
+            FieldType o1 = customField1.getFieldType();
+            FieldType o2 = customField2.getFieldType();
+            String className1 = o1 == null ? "Unknown" : o1.getClass().getSimpleName();
+            String className2 = o2 == null ? "Unknown" : o2.getClass().getSimpleName();
+            String fieldName1 = o1 == null ? "Unknown" : o1.getName();
+            String fieldName2 = o2 == null ? "Unknown" : o2.getName();
+            String name1 = className1 + "." + fieldName1 + " " + customField1.getAlias();
+            String name2 = className2 + "." + fieldName2 + " " + customField2.getAlias();
+            return name1.compareTo(name2);
+         }
+      });
+
+      for (CustomField field : allCustomFields)
+      {
+         if (!field.getLookupTable().isEmpty())
+         {
+            if (outlineCodes == null)
+            {
+               outlineCodes = m_factory.createProjectOutlineCodes();
+               project.setOutlineCodes(outlineCodes);
+            }
+
+            Project.OutlineCodes.OutlineCode outlineCode = m_factory.createProjectOutlineCodesOutlineCode();
+            outlineCodes.getOutlineCode().add(outlineCode);
+            writeOutlineCode(outlineCode, field);
+         }
+      }
+   }
+
+   /**
+    * Write a single outline code or custom field.
+    * 
+    * @param outlineCode outline codes root node
+    * @param field custom field
+    */
+   private void writeOutlineCode(Project.OutlineCodes.OutlineCode outlineCode, CustomField field)
+   {
+      //
+      // Header details
+      //
+      CustomFieldLookupTable table = field.getLookupTable();
+      outlineCode.setGuid(table.getGUID());
+      outlineCode.setEnterprise(Boolean.valueOf(table.getEnterprise()));
+      outlineCode.setShowIndent(Boolean.valueOf(table.getShowIndent()));
+      outlineCode.setResourceSubstitutionEnabled(Boolean.valueOf(table.getResourceSubstitutionEnabled()));
+      outlineCode.setLeafOnly(Boolean.valueOf(table.getLeafOnly()));
+      outlineCode.setAllLevelsRequired(Boolean.valueOf(table.getAllLevelsRequired()));
+      outlineCode.setOnlyTableValuesAllowed(Boolean.valueOf(table.getOnlyTableValuesAllowed()));
+
+      //
+      // Masks
+      //
+      outlineCode.setMasks(m_factory.createProjectOutlineCodesOutlineCodeMasks());
+      if (field.getMasks().isEmpty())
+      {
+         CustomFieldValueDataType type = table.get(0).getType();
+         if (type == null)
+         {
+            type = CustomFieldValueDataType.TEXT;
+         }
+         CustomFieldValueMask item = new CustomFieldValueMask(0, 1, ".", type);
+         writeMask(outlineCode, item);
+      }
+      else
+      {
+         for (CustomFieldValueMask item : field.getMasks())
+         {
+            writeMask(outlineCode, item);
+         }
+      }
+
+      //
+      // Values
+      //
+      Project.OutlineCodes.OutlineCode.Values values = m_factory.createProjectOutlineCodesOutlineCodeValues();
+      outlineCode.setValues(values);
+
+      for (CustomFieldValueItem item : table)
+      {
+         Project.OutlineCodes.OutlineCode.Values.Value value = m_factory.createProjectOutlineCodesOutlineCodeValuesValue();
+         values.getValue().add(value);
+         writeOutlineCodeValue(value, item);
+      }
+   }
+
+   /**
+    * Write an outline code value.
+    * 
+    * @param value parent node
+    * @param item custom field item
+    */
+   private void writeOutlineCodeValue(Project.OutlineCodes.OutlineCode.Values.Value value, CustomFieldValueItem item)
+   {
+      CustomFieldValueDataType type = item.getType();
+      if (type == null)
+      {
+         type = CustomFieldValueDataType.TEXT;
+      }
+      value.setDescription(item.getDescription());
+      value.setFieldGUID(item.getGUID());
+      value.setIsCollapsed(Boolean.valueOf(item.getCollapsed()));
+      value.setParentValueID(NumberHelper.getBigInteger(item.getParent()));
+      value.setType(BigInteger.valueOf(type.getValue()));
+      value.setValueID(NumberHelper.getBigInteger(item.getUniqueID()));
+      value.setValue(DatatypeConverter.printOutlineCodeValue(item.getValue(), type.getDataType()));
+   }
+
+   /**
+    * Write an outline code mask element.
+    * 
+    * @param outlineCode parent node
+    * @param item mask element
+    */
+   private void writeMask(Project.OutlineCodes.OutlineCode outlineCode, CustomFieldValueMask item)
+   {
+      Project.OutlineCodes.OutlineCode.Masks.Mask mask = m_factory.createProjectOutlineCodesOutlineCodeMasksMask();
+      outlineCode.getMasks().getMask().add(mask);
+
+      mask.setLength(BigInteger.valueOf(item.getLength()));
+      mask.setLevel(BigInteger.valueOf(item.getLevel()));
+      mask.setSeparator(item.getSeparator());
+      mask.setType(BigInteger.valueOf(item.getType().getMaskValue()));
    }
 
    /**
@@ -1493,13 +1640,13 @@ public final class MSPDIWriter extends AbstractProjectWriter
             dummy.setWork(duration);
             dummy.setActualWork(Duration.getInstance(actualWork, durationUnits));
             dummy.setRemainingWork(Duration.getInstance(remainingWork, durationUnits));
-            
+
             // Without this, MS Project will mark a 100% complete milestone as 99% complete
             if (percentComplete == 100 && duration.getDuration() == 0)
-            {              
+            {
                dummy.setActualFinish(task.getActualStart());
             }
-            
+
             list.add(writeAssignment(dummy));
          }
       }