+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 7.9.8 (git master)
+* Added support for reading and writing outline code/custom field lookup tables for MSPDI files.
 
 ## 7.9.7 (25/11/2019)
 * Round percent complete values read from Asta files to two decimal places to avoid values like 99.9999999%.